requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.42.2
colorama==0.4.6
coverage==7.6.10
Django==5.1.5
django-cloudinary-storage==0.3.0
django-extensions==3.2.3
django-webpack-loader==3.1.1
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.0.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
whitenoise==6.9.0
gunicorn==21.2.0
